+++ Library/include/CSP/Systems/Spaces/SpaceSystem.h
@@ -247,13 +247,13 @@ public:
     /// @param SpaceId csp::common::String : Space ID to associate the Site information with
     /// @param SiteInfo Site : Site information to be added
     /// @param Callback SiteResultCallback : callback when asynchronous task finishes
-    CSP_ASYNC_RESULT void AddSiteInfo(const csp::common::String& SpaceId, Site& SiteInfo, SiteResultCallback Callback);
+    CSP_ASYNC_RESULT void AddSiteInfo(const csp::common::String& SpaceId, Site SiteInfo, SiteResultCallback Callback);
 
     /// @brief Removes the Site information from the Space.
     /// @param SpaceId csp::common::String : Space ID for which to remove the associated Site information
     /// @param SiteInfo Site : Site information to be removed
     /// @param Callback NullResultCallback : callback when asynchronous task
-    CSP_ASYNC_RESULT void RemoveSiteInfo(const csp::common::String& SpaceId, Site& SiteInfo, NullResultCallback Callback);
+    CSP_ASYNC_RESULT void RemoveSiteInfo(const csp::common::String& SpaceId, Site SiteInfo, NullResultCallback Callback);
 
     /// @brief Retrieves the Sites information associated with a Space.
     /// @param SpaceId csp::common::String : Space ID to be queried for Site information